Declare latency histograms DOUBLE so fractional ms stop warning (Fixes #3259)
Both latency histograms (llxprt_code.api.request.latency and
llxprt_code.tool.call.latency) were declared ValueType.INT while every
recorded duration is a performance.now() delta in fractional
milliseconds. The OTel SDK warns on each non-integer recording,
truncates, and records; with DiagConsoleLogger wired at INFO the
warning prints on nearly every API request. In non-interactive runs
(-p) that line interleaves with the response on stdout and breaks
piped consumers.

Latency in milliseconds is naturally fractional and OTel duration
metrics are conventionally double, so the two histograms are now
ValueType.DOUBLE. Durations keep sub-millisecond precision instead of
being silently truncated. Counters stay INT: they only ever record
literal 1 or integer token counts.

A real-SDK behavioral test locks the contract: it registers a real
MeterProvider with an in-memory reader, records fractional durations
through the public record functions, and asserts the exported
histogram sums keep the exact fractional values with no INT
truncation diagnostic captured. Verified RED against the old INT
declarations (sums truncated to 823 and 17) and GREEN after the
change; smoke run shows zero warnings.

+++ packages/telemetry/src/telemetry/metrics.ts
@@ -67,7 +67,7 @@ export function initializeMetrics(config: TelemetryConfig): void {
   toolCallLatencyHistogram = meter.createHistogram(METRIC_TOOL_CALL_LATENCY, {
     description: 'Latency of tool calls in milliseconds.',
     unit: 'ms',
-    valueType: ValueType.INT,
+    valueType: ValueType.DOUBLE,
   });
   apiRequestCounter = meter.createCounter(METRIC_API_REQUEST_COUNT, {
     description: 'Counts API requests, tagged by model and status.',
@@ -78,7 +78,7 @@ export function initializeMetrics(config: TelemetryConfig): void {
     {
       description: 'Latency of API requests in milliseconds.',
       unit: 'ms',
-      valueType: ValueType.INT,
+      valueType: ValueType.DOUBLE,
     },
   );
   tokenUsageCounter = meter.createCounter(METRIC_TOKEN_USAGE, {
